Link libX11 dynamically
Building libX11 statically causes problems with libpthread, however
the version of libX11 that ships with el5 is too old to build libXext.
So we have to build a dynamic version of libX11 and link against that
while building.  The resulting binary will actually be linked against
the system libX11, which should be ABI compatible as far as Xvnc is
concerned.