contrib: Enable ubuntu-xenial debian source package to build on xenial.
The package is based on a copy of the ubuntu-trusty debian package. The
following modifications have been made:

https://github.com/TigerVNC/tigervnc/archive/v1.6.0.tar.gz can be used
unmodified as the .orig tarball for the debian package. Therefor the
get-orig-source.sh helper is no longer there.

xorg-server-source has been added to the build dependencies, but since
tigervnc 1.6.0 does not support xorg 1.18, usage of that package is
commented out in debian/rules. A copy of xorg-server-1.17.4.tar.bz2 has to
be provided instead, or it is fetched on demand by debian/rules from
http://xorg.freedesktop.org/releases/individual/xserver/. This
inconsistency can be removed as soon as tigervnc supports xorg 1.18.

Since the xorg sources are kept separate from the .orig tarball, the two
small patches which where in the trusty quilt series are now applied after
the xorg sources have been extracted. There is no quilt series anymore at
this point.

Fltk 1.3.3 is available in the xenial repositories and it is used to build
xtigervncviewer. No need to get the fltk sources. patch_fltk.sh is gone.

Fixed some dependencies to reflect changes between trusty and xenial.

Note: The packages have been built in a clean xenial amd64 debootstrap
chroot (2016-04-07) and tried.

Note: Some dependencies (including libfltk1.3 and xorg-server-source) are
found in the xenial universe repository only.

Note: Using the latest sources on the master branch for the .orig tarball
works as well. But during binary package installation there is some trouble
(which can be worked around using dpkg -i --force-overwrite):

dpkg: error processing archive
../xtigervncviewer_1.6.0master20160407_amd64.deb (--install): trying to
overwrite '/usr/share/applications/vncviewer.desktop', which is also in
package tigervncserver 1.6.0master20160407