contrib: Add ubuntu-xenial debian source package to build on xenial.
The package is based on the contrib ubuntu-trusty debian package. The
following modifications have been made:

The current tigervnc sources can be used unmodified as the .orig
tarball for the debian package, so get-orig-source.sh is gone.

The ubuntu xorg-server-sources package provides the xorg sources used
to build tigervnc. The two small patches which where in the trusty
quilt series are now applied to the xorg sources after they have been
extracted. There is no quilt series anymore at this point.

Fltk 1.3.3 is available in the xenial repositories and it is used to build
xtigervncviewer. No need to get the fltk sources. patch_fltk.sh is gone.

Fixed some dependencies to reflect changes between trusty and xenial.

Removed configure options not recognized by configure.

The install target is not called a second time for the viewer package,
but the viewer files are moved from the server installation. (Eliminate
file conflicts during binary package installation.)

Note: Some dependencies (including libfltk1.3 and xorg-server-source) are
found in the xenial universe repository only.