Use cmake TIMESTAMP function
because it is not only platform independent
but also allows to override the build date
This helps to make tigervnc builds reproducible

See https://reproducible-builds.org/ for why this is good

Also adds UTC flag, to be independent of timezone.
Also changes JAVA_DATE format to ISO-8601 date format.

Requires cmake-2.8.11+ from 2013