Fix "vncserver -list" to have deterministic order and correct filter.
When listing sessions, use lexicographical order insead of directory order.

The filtering regular expression is vulnerable to misinterpretation, and
include a strange reference to ` $\ ` that is likely a typo:

       if ($file =~ /$host:(\d+)$\.pid/) {

It only works because ` $\ ` is normally empty, so it is acting like:

       if ($file =~ /$host:(\d+).pid/) {

This pattern is vulnerable if `$host` contains any regular expression
characters, including `.`, and also the `.` before pid is not literal.

This has been corrected to:

       if ($file =~ /^\Q$host\E:(\d+)\.pid$/) {