Fix "vncserver" display check to better identify available display numbers.
There are a number of situations than can occur when the X
server is not shut down cleanly. These situations can leave
files around that "vncserver" has previously mis-identified
as evidence that the display number is still in use. These
changes reduce the occurrences of these false positives,
ensuring that "vncserver" handles these situations gracefully.

Instead of checking for existence of `/tmp/.X<n>-lock`, the
code will now extract the PID from this file and confirm that
a process exists with the same PID. This will eliminate
false positives in the case that this file references a PID
that no longer exists. The Xorg server does not have a problem
with fixing this file when it next starts up. It is only
important to avoid using the port if it is still in use.

Instead of checking for existence of `/tmp/.X11-unix/X<n>`, the
code will now attempt to connect to the socket to confirm that
there is a server process listening on this UNIX domain socket.
The Xorg server does not have a problem with fixing this file
when it next starts up. It is only important to avoid using
the UNIX domain socket if it is still in use.

The ordering of the code has changed, such that that check for
a lock file happens first. This should ensure that the fast
path does not require to bind() or connect() to any sockets
which can both have side effects. The slow path will still
do all checks, so there is no risk to this change.

The check for existence of `/usr/spool/sockets/X11/<n>` has
been removed. This file is only relevant on HP-UX, and
TigerVNC dropped support for HP-UX in commit 31e5aa3ddd2.

The section of code updated now requires Perl 5. This
removes legacy cruft that should not be necessary on any
supported platforms.