EL 7 tigervnc.spec updates
Add missing build requirements: libpng-devel, zlib-devel

Remove obsolete build requirements: libXfont-devel

Fix unexpected rpm macro expansion `%{VERSION}` which is supposed
to be passed through as literal to `rpm -q --qf`. This hasn't
caused problems before, but it was found to cause problems when
built from EL 8 when it expanded to "1.10.1".