Call correct wrappers for clipboard handling
We incorrectly called the underlying functions instead of the safe
wrappers for the new clipboard functions. This had the effect of a)
crashing the entire server if one of these functions failed, and b) not
respecting the settings disabling the clipboard.

(cherry picked from commit 331a27addf46d39635fb4d195ae2f94058689832)