Calculate overlay position from viewport
Previously, it was assumed that all monitors (and especially the primary
monitor with index 0) was included in the viewport when calculating the
position of the overlay.

Because that might not always be true when having fullscreen mode with a
narrower set of monitors, this commit does so the overlay is positioned
correctly, based on what the viewport actually is, instead of what we
expect it to be.