TEMP: Use event taps for keyboard grab on macOS
These are the slightly more proper way to grab the keyboard and macOS
and is what similar applications do. It avoids a lot of issues we have,
e.g., problems with multiple monitors.

Unfortunately we need to have the user explicitly approve this (which
really is a good thing, security wise), and Apple have chosen to mark
this feature as only for accessibility.