Synchronously retry X11 keyboard grab
Make this less complex by removing the timer and retrying things
synchronously. This can freeze the UI, but we give up after half a
second so it should hopefully not be noticable.

The advantage is that we can directly determine if we succeeded or
failed grabbing the keyboard. The previous code could in theory continue
retrying forever.