Change to modern layout in options dialog
Classical tabs are very dated. They are also a practical problem as you
get very limited in the numbers of tabs we can have, and how long the
text can be on them.

Switch to one popular modern model with a list on the left instead where
pages can be selected.